feat(workbench): split and refine git panels
