+++ packages/workbench-ui/src/lib/panel/PanelToolbarButton.svelte
@@ -1,6 +1,9 @@
 <script lang="ts">
 import type { Component, Snippet } from "svelte";
-import { Button } from "@nervekit/ui-kit/components/ui/button";
+import {
+  Button,
+  type ButtonVariant,
+} from "@nervekit/ui-kit/components/ui/button";
 import { Spinner } from "@nervekit/ui-kit/components/ui/spinner";
 import { cn } from "@nervekit/ui-kit/core/utils";
 
@@ -8,6 +11,7 @@ let {
   icon: Icon,
   label,
   title,
+  variant = "ghost",
   active = false,
   disabled = false,
   loading = false,
@@ -20,6 +24,7 @@ let {
   /** Accessible name; also the tooltip when `title` is omitted. */
   label: string;
   title?: string;
+  variant?: ButtonVariant;
   active?: boolean;
   disabled?: boolean;
   /** Swaps the icon for a spinner while an action is in flight. */
@@ -34,7 +39,7 @@ let {
 </script>
 
 <Button
-  variant="ghost"
+  {variant}
   size={showLabel ? "xs" : "icon-xs"}
   class={cn("shrink-0", className)}
   ariaLabel={label}
